Remove dep to Ganache for dist
