+++ test/contracts/boilerplate/base.js
@@ -10,7 +10,8 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinonChai from 'sinon-chai';
-import {createWeb3, deployContract} from '../../../src/utils/web3_tools';
+import {deployContract} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 import chaiEmitEvents from '../../helpers/chaiEmitEvents';
 import observeBalanceChange from '../../helpers/web3BalanceObserver';
 
@@ -40,7 +41,7 @@ describe('Base Contract', () => {
 
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     [deployer, catalogue, storageCatalogue, migrator, fundsMigrationTarget] = await web3.eth.getAccounts();
 
     head = await deployContract(web3, HeadJson, [deployer], {from: deployer});
