+++ test/contracts/boilerplate/context.js
@@ -10,7 +10,8 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinonChai from 'sinon-chai';
-import {createWeb3, deployContract} from '../../../src/utils/web3_tools';
+import {deployContract} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 import ContextJson from '../../../src/contracts/Context.json';
 
 chai.use(sinonChai);
@@ -34,7 +35,7 @@ describe('Context Contract', () => {
   const versionTag = async (contract) => contract.methods.versionTag().call();
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     [deployer, trustedAddress, catalogueAddress, storageCatalogueAddress, untrustedAddress] = await web3.eth.getAccounts();
   });
