+++ test/contracts/boilerplate/head.js
@@ -10,7 +10,8 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinonChai from 'sinon-chai';
-import {createWeb3, deployContract} from '../../../src/utils/web3_tools';
+import {deployContract} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 import HeadJson from '../../../src/contracts/Head.json';
 
 chai.use(sinonChai);
@@ -36,7 +37,7 @@ describe('Head Contract', () => {
   const transferOwnership = async (contract, sender, newOwner) => contract.methods.transferOwnership(newOwner).send({...standardOptions, from: sender});
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     [deployAddress, ownerAddress, nonOwnerAddress, otherAddress] = await web3.eth.getAccounts();
   });
