+++ test/contracts/consensus/constructor_ownable.js
@@ -9,7 +9,8 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import {createWeb3, deployContract, makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {deployContract, makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 import chaiEmitEvents from '../../helpers/chaiEmitEvents';
 import ConstructorOwnableJson from '../../../src/contracts/ConstructorOwnable.json';
 
@@ -35,7 +36,7 @@ describe('ConstructorOwnable base contract', () => {
   let snapshotId;
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     [deployer, owner, otherUser, newOwner] = await web3.eth.getAccounts();
     contract = await deploy(web3, deployer, owner);
   });
