+++ test/contracts/front/payouts.js
@@ -13,7 +13,8 @@ import sinonChai from 'sinon-chai';
 import deploy from '../../helpers/deploy';
 import TimeMockJson from '../../../src/contracts/TimeMock.json';
 import observeBalanceChange from '../../helpers/web3BalanceObserver';
-import {createWeb3, makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 import {PAYOUT_PERIOD_UNIT} from '../../helpers/consts';
 
 chai.use(sinonChai);
@@ -46,7 +47,7 @@ describe('Payouts Contract', () => {
   const expectBalanceChange = async (account, amount, codeBlock) => expect((await observeBalanceChange(web3, account, codeBlock)).toString()).to.eq(amount);
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     [validUser, beneficiary, withdrawTarget, otherUser] = await web3.eth.getAccounts();
     ({payoutsStore, payouts, time} = await deploy({
       web3,
