+++ test/contracts/pool/pools_nodes_manager.js
@@ -9,7 +9,8 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 
 import {expect, assert} from '../../helpers/chaiPreconf';
 import deploy from '../../helpers/deploy';
-import {createWeb3, makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 import {ROLE_CODES, ZERO_ADDRESS, APOLLO_DEPOSIT, ATLAS1_STAKE} from '../../../src/constants';
 import {utils} from 'web3';
 
@@ -56,7 +57,7 @@ describe('PoolsNodesManager Contract', () => {
   const addNodeRequestResolved = (requestId, status, senderAddress = owner) => poolsNodesManager.methods.addNodeRequestResolved(requestId, status).send({from: senderAddress});
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     web3.eth.handleRevert = true;
     [owner, initialApollo, pool, pool2, node, user] = await web3.eth.getAccounts();
