+++ test/contracts/storage/payouts_store.js
@@ -12,7 +12,8 @@ import chaiAsPromised from 'chai-as-promised';
 import sinonChai from 'sinon-chai';
 import deploy from '../../helpers/deploy';
 import observeBalanceChange from '../../helpers/web3BalanceObserver';
-import {createWeb3, makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 
 chai.use(sinonChai);
 chai.use(chaiAsPromised);
@@ -38,7 +39,7 @@ describe('PayoutsStore Contract', () => {
   const expectBalanceChange = async (account, amount, codeBlock) => expect((await observeBalanceChange(web3, account, codeBlock)).toString()).to.eq(amount);
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     [validUser, beneficiary, withdrawTarget, otherUser] = await web3.eth.getAccounts();
     ({payoutsStore} = await deploy({
       web3,
