+++ test/contracts/storage/pools_store.js
@@ -9,7 +9,8 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 
 import {expect, assert} from '../../helpers/chaiPreconf';
 import deploy from '../../helpers/deploy';
-import {createWeb3, makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {makeSnapshot, restoreSnapshot} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 import {ZERO_ADDRESS} from '../../../src/constants';
 
 describe('PoolsStore Contract', () => {
@@ -29,7 +30,7 @@ describe('PoolsStore Contract', () => {
   const getPools = (from, to, senderAddress = owner) => poolsStore.methods.getPools(from, to).call({from: senderAddress});
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     web3.eth.handleRevert = true;
     [owner, pool] = await web3.eth.getAccounts();
