+++ test/contracts/storage/sheltering_transfers_store.js
@@ -9,7 +9,8 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import {createWeb3, makeSnapshot, restoreSnapshot, utils} from '../../../src/utils/web3_tools';
+import {makeSnapshot, restoreSnapshot, utils} from '../../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../../utils/web3_tools';
 import deploy from '../../helpers/deploy';
 
 chai.use(chaiAsPromised);
@@ -36,7 +37,7 @@ describe('ShelteringTransfersStore Contract', () => {
     shelteringTransfersStore.methods.getTransferId(donorId, bundleId).call();
 
   before(async () => {
-    web3 = await createWeb3();
+    web3 = await createWeb3Ganache();
     [contextAddress, donorAddress, otherAddress] = await web3.eth.getAccounts();
     ({shelteringTransfersStore} = await deploy({
       web3,
