+++ test/helpers/deploy.js
@@ -10,10 +10,11 @@ This Source Code Form is “Incompatible With Secondary Licenses”, as defined 
 import MockContextJson from '../../src/contracts/MockContext.json';
 import Deployer from '../../src/deployer';
 import contractJsons from '../../src/contract_jsons';
-import {createWeb3, getDefaultAddress} from '../../src/utils/web3_tools';
+import {getDefaultAddress} from '../../src/utils/web3_tools';
+import {createWeb3Ganache} from '../utils/web3_tools';
 
 const deploy = async (options = {}) => {
-  const web3 = options.web3 || await createWeb3();
+  const web3 = options.web3 || await createWeb3Ganache();
   const sender = options.sender || getDefaultAddress(web3);
   const skipDeployment = [];
   const alreadyDeployed = {};
